перевод логики с базы данных на сервер.
