+++ app/src/main/java/ru/miel/data/network/api/CandidatesApi.kt
@@ -1,4 +1,4 @@
-package ru.miel.data.api
+package ru.miel.data.network.api
 
 import retrofit2.http.Body
 import retrofit2.http.DELETE
@@ -6,9 +6,9 @@ import retrofit2.http.GET
 import retrofit2.http.Header
 import retrofit2.http.POST
 import retrofit2.http.Path
-import ru.miel.data.dto.candidates.request.SetFlagByIdRequest
-import ru.miel.data.dto.candidates.response.CandidatesByFlagResponse
-import ru.miel.data.dto.candidates.response.CandidatesResponse
+import ru.miel.data.network.dto.candidates.request.SetFlagByIdRequest
+import ru.miel.data.network.dto.candidates.response.CandidatesByFlagResponse
+import ru.miel.data.network.dto.candidates.response.CandidatesResponse
 
 interface CandidatesApi {
     @GET("/api/supervisor/candidates/")
