+++ app/src/main/java/ru/miel/data/network/api/UserApi.kt
@@ -1,12 +1,12 @@
-package ru.miel.data.api
+package ru.miel.data.network.api
 
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.Header
 import retrofit2.http.POST
-import ru.miel.data.dto.user.request.AuthUserRequest
-import ru.miel.data.dto.user.response.AuthUserResponse
-import ru.miel.data.dto.user.response.InfoUserResponse
+import ru.miel.data.network.dto.user.request.AuthUserRequest
+import ru.miel.data.network.dto.user.response.AuthUserResponse
+import ru.miel.data.network.dto.user.response.InfoUserResponse
 
 interface UserApi {
     @POST("/api/login/")
