+++ app/src/main/java/ru/miel/data/network/dto/candidates/request/SetFlagByIdRequest.kt
@@ -1,4 +1,4 @@
-package ru.miel.data.dto.candidates.request
+package ru.miel.data.network.dto.candidates.request
 
 data class SetFlagByIdRequest(
     val candidate: Int,
