+++ app/src/main/java/ru/miel/data/network/dto/candidates/response/CandidatesByFlagResponse.kt
@@ -1,6 +1,6 @@
-package ru.miel.data.dto.candidates.response
+package ru.miel.data.network.dto.candidates.response
 
-import ru.miel.data.dto.candidates.CandidateDto
+import ru.miel.data.network.dto.candidates.CandidateDto
 
 data class CandidatesByFlagResponse(
     val id: Int,
