+++ app/src/main/java/ru/miel/data/network/dto/candidates/response/CandidatesResponse.kt
@@ -1,7 +1,8 @@
-package ru.miel.data.dto.candidates
+package ru.miel.data.network.dto.candidates.response
 
-data class CandidateDto(
+data class CandidatesResponse(
     val id: Int,
+    val is_active: Boolean,
     val name: String,
     val surname: String,
     val patronymic: String,
@@ -10,12 +11,16 @@ data class CandidateDto(
     val photo: String?,
     val country: String,
     val city: String,
+    val email: String,
     val resume: String,
+    val is_free: Boolean,
     val course_rieltor_join: String,
     val basic_legal_course: String,
     val course_mortgage: String,
     val course_taxation: String,
     val completed_objects: Int,
     val clients: Int,
+    val created_at: String,
     val updated_at: String,
+    val office: String?,
 )
