+++ app/src/main/java/ru/miel/di/modules/DataModule.kt
@@ -3,8 +3,8 @@ package ru.miel.di.modules
 import android.content.SharedPreferences
 import dagger.Module
 import dagger.Provides
-import ru.miel.data.api.CandidatesApi
-import ru.miel.data.api.UserApi
+import ru.miel.data.network.api.CandidatesApi
+import ru.miel.data.network.api.UserApi
 import ru.miel.data.dbo.dao.CandidatesDao
 import ru.miel.data.dbo.dao.QuotesDao
 import ru.miel.data.repository.CandidatesRepository
