+++ app/src/main/java/ru/miel/view/showcase/ShowcaseViewModel.kt
@@ -8,6 +8,7 @@ import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.launch
 import ru.miel.domain.models.Candidates
+import ru.miel.domain.models.CandidatesFromApi
 import ru.miel.domain.usecase.candidates.GetCandidatesApiUseCase
 import ru.miel.domain.usecase.candidates.GetCandidatesDbUseCase
 import ru.miel.domain.usecase.candidates.GetQuotesByNowDbUseCase
@@ -25,18 +26,19 @@ class ShowcaseViewModel(
 ) :
     ViewModel() {
     /***Кандидаты с сервера*/
-//    private var _candidates = MutableSharedFlow<List<CandidatesFromApi>>()
-//    val candidates: SharedFlow<List<CandidatesFromApi>>
-//        get() = _candidates.asSharedFlow()
+    private var _candidates = MutableSharedFlow<List<CandidatesFromApi>>()
+    val candidates: SharedFlow<List<CandidatesFromApi>>
+        get() = _candidates.asSharedFlow()
     /****/
 
     init {
         getQuotes()
     }
 
-    private var _candidates = MutableSharedFlow<List<Candidates>>()
-    val candidates: SharedFlow<List<Candidates>>
-        get() = _candidates.asSharedFlow()
+    /***Кандидаты с BD, мок-данные*/
+//    private var _candidates = MutableSharedFlow<List<Candidates>>()
+//    val candidates: SharedFlow<List<Candidates>>
+//        get() = _candidates.asSharedFlow()
 
     private var _quotes = MutableSharedFlow<Int>()
     val quotes: SharedFlow<Int>
@@ -70,8 +72,8 @@ class ShowcaseViewModel(
 
     fun getCandidates() {
         viewModelScope.launch {
-//            val result = getCandidatesApiUseCase()
-            val result = getCandidatesDbUseCase()
+            val result = getCandidatesApiUseCase()
+//            val result = getCandidatesDbUseCase()
             _candidates.emit(result)
         }
     }
